+++ ecorea/src/main/java/com/project/ecorea/service/CartService.java
@@ -18,11 +18,6 @@ public class CartService {
 	private CartDao cartDao;
 	private ProductDao productDao;
 	
-	/* 장바구니 담기 */
-	public Boolean add(Integer pno, Integer count, String loginId) {
-		
-	}
-	
 	// 장바구니 목록 출력
 	public List<CartDto.CartList> readCart(String memberId) {
 		return cartDao.findByMemberId(memberId);
