+++ ecorea/src/main/java/com/project/ecorea/service/JumunService.java
@@ -4,6 +4,9 @@ import java.util.*;
 
 import org.springframework.stereotype.*;
 
+import com.example.demo.dto.OrderDto;
+import com.example.demo.entity.OrderItem;
+import com.example.demo.entity.Product;
 import com.project.ecorea.dao.*;
 import com.project.ecorea.dto.*;
 import com.project.ecorea.dto.CartDto.*;
@@ -19,6 +22,8 @@ public class JumunService {
 	private UserDao memberDao;
 	private AddressDao addressDao;
 	
+	/* 상품 -> 바로 구매 */
+
 
 	public JumunDto.JumunPreview jumunList(List<Params> list, String memberId) {	
 		Integer totalPrice = 0;
