+++ ecorea/src/main/resources/application.properties
@@ -2,7 +2,7 @@ server.port=8081
 
 spring.datasource.driverClassName=net.sf.log4jdbc.sql.jdbcapi.DriverSpy
 spring.datasource.url=jdbc:log4jdbc:oracle:thin:@localhost:1521:XE
-spring.datasource.username=ecorea
+spring.datasource.username=c##ecorea
 spring.datasource.password=1234
