+++ ecorea/src/main/resources/templates/order/cart.html
@@ -78,8 +78,7 @@
 			}).fail(()=>alert("최소 구매 수량은 1개 입니다."));
 		}
 		
-		
-		// 상품 한개 삭제
+		// 상품 한 개 삭제
 		function deleteOne() {
 			const result = confirm("상품을 장바구니에서 삭제하시겠습니까?");
 			if(result==true) {
@@ -88,7 +87,6 @@
 				$form.submit();
 			}
 		}
-		
 		
 		// 상품 전체 삭제
 		function deleteAll() {
