+++ ecorea/src/main/resources/templates/product/member/productDetail.html
@@ -83,12 +83,31 @@
   		}
   	}
   	
+  	/* 바로 구매 버튼 */
+  	function toOrder() {
+		const $form = $('<form>').attr('method', 'post').attr('action', '/order/preview/one').appendTo($('body'));
+		$('<input>').attr('type', 'hidden').attr('name','pno').val(product.pno).appendTo($form);
+		$('<input>').attr('type', 'hidden').attr('name','count').val($('#p_num1').val()).appendTo($form);
+		$form.submit();
+  	}
+  	
+  	/* 장바구니 버튼 */
+  	function toCart() {
+		const $form = $('<form>').attr('method', 'post').attr('action', '/order/cart/add').appendTo($('body'));
+		$('<input>').attr('type', 'hidden').attr('name','pno').val(product.pno).appendTo($form);
+		$('<input>').attr('type', 'hidden').attr('name','count').val($('#p_num1').val()).appendTo($form);
+		$form.submit();
+  	}
+  	
   	$(document).ready(function() {
   		reviewList();
   		qnaQList();
   		$('#plus').click(plusCnt);
   		$('#minus').click(minusCnt);
+  		$('#order').click(toOrder);
+  		$('#cart').click(toCart);
   	})
+  	
   </script>
   <title>상품 상세페이지</title>
   <style>
@@ -266,9 +285,9 @@ img {vertical-align: middle; height: 300px;}
         <option>3</option>
       </select><br>
       <span>총 합계 : </span><span><span id='totalPrice'></span>&nbsp;원</span><span> (<span id='totalCnt'></span>&nbsp;개)</span><br><br>
-      <input type="submit" class="btn btn-primary" value="바로구매" formaction="../../order/pay.html">
-      <input type="submit" class="btn btn-primary" value="장바구니 담기" formaction="../../order/cart.html">
-      <input type="submit" class="btn btn-primary" value="관심상품 등록" formaction="#">
+      <button type="button" class="btn btn-primary" id="order" value="바로구매"></button>
+      <button type="button" class="btn btn-primary" id="cart" value="장바구니 담기"></button>
+      <button type="button" class="btn btn-primary" value="관심상품 등록" formaction="#"></button>
     </div>
       <div class="inside-product topnav topnav-right">
       <a href="#content1">상품 상세정보</a>
